chore: split up test into package-files
