+++ package.json
@@ -16,7 +16,7 @@
 		"ard:publishers": "node -r dotenv/config scripts/ard/publishers.js",
 		"ard:pub-sort": "node -r dotenv/config scripts/ard/sortPubByExtId.js",
 		"lint": "eslint .",
-		"test": "mocha test/test.js -r dotenv/config",
+		"test": "mocha tests/**.js -r dotenv/config",
 		"reinstall": "rm -rf node_modules && rm yarn.lock && yarn"
 	},
 	"author": "SWR Audio Lab <lab@swr.de>",
